Moved calculator into seperate module, freshened it up and implemented month changer
